feat: support node 22
This snapshot takes the suggestion from
https://github.com/v8/v8/commit/e48c47268c53499814c88c35064d564dbb92e6fd
(where ALL_CAN_READ was removed) and switches over to using
SetNativeDataProperty for the accessors.

Unfortunately, that method was introduced some time in the V8 9.x
series, which node didn't adopt until sometime in the 16.x series after
16.14.0, and the signature changed in a few ways between its
introduction and the aforementioned removal of ALL_CAN_READ (which
appears to have landed in v8 version 12.1.0).

So this represents a conservative position that defers uptake of the new
method until very recent v8 (and thus node) versions, but that kept the
`#ifdef`ing to a minimum in a way that will hopefully be easy to remove
as support for old versions drops off.